+++ NEWS.md
@@ -3,6 +3,8 @@
 ## Changes in version 2020.1.9000
 
 - In `record_sdd()`, date is now recorded in GMT timezone and with microseconds.
+- An `init()` function is added to set globally `MONGO_URL` and `MONGO_BASE`
+  environment variables for lerarndown Shiny and learnr applications.
 
 ## Changes in version 2020.0.9000