+++ NEWS.md
@@ -1,5 +1,11 @@
 # BioDataScience News
 
+## Changes in version 2020.0.9000
+
+- In `record_sdd()`, date is now recorded in GMT timezone and with microseconds.
+- An `init()` function is added to set globally `MONGO_URL` and `MONGO_BASE`
+  environment variables for lerarndown Shiny and learnr applications.
+
 ## Changes in version 2019.6.0
 
 - correction tutorial 06b_recombinaison